Refactor tokenize() : argument change to path.
