+++ inc/file.h
@@ -1,0 +1,7 @@
+#ifndef FILE_INC_H
+#define FILE_INC_H
+
+#include "mcc.h"
+char* read_file(char* path);
+
+#endif
